requirements.txt
# Core dependencies
pydantic>=2.0.0
numpy==1.26.3

# LangChain ecosystem
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-neo4j

# LLM providers
ollama

# Database & vector stores
neo4j
chromadb

# PDF processing (Streamlit Cloud compatible)
unstructured[pdf]
pdfminer.six
pymupdf
pypdf

# Image processing & OCR (Streamlit Cloud compatible)
opencv-python-headless==4.8.0.76
pytesseract

# Data processing
json_repair

# Reference management
pyzotero

# Visualization (for Streamlit app)
streamlit
pandas
networkx
pyvis
plotly

# Configuration management
python-dotenv